// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const Role = {
  Chef: "Chef",
  Waiter: "Waiter",
  Bartender: "Bartender",
  Manager: "Manager",
  Cook: "Cook",
  Kitchen_Hand: "Kitchen_Hand",
} as const;

export type Role = (typeof Role)[keyof typeof Role];

export const LeaveStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type LeaveStatus = (typeof LeaveStatus)[keyof typeof LeaveStatus];

export const WeekDay = {
  MON: "MON",
  TUE: "TUE",
  WED: "WED",
  THU: "THU",
  FRI: "FRI",
  SAT: "SAT",
  SUN: "SUN",
} as const;

export type WeekDay = (typeof WeekDay)[keyof typeof WeekDay];

export const BookingStatus = {
  PENDING: "PENDING",
  CONFIRMED: "CONFIRMED",
  SEATED: "SEATED",
  COMPLETED: "COMPLETED",
  CANCELLED: "CANCELLED",
  NO_SHOW: "NO_SHOW",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const MenuCategory = {
  APPETIZER: "APPETIZER",
  MAIN_COURSE: "MAIN_COURSE",
  DESSERT: "DESSERT",
  BEVERAGE: "BEVERAGE",
  SIDE: "SIDE",
  SALAD: "SALAD",
} as const;

export type MenuCategory = (typeof MenuCategory)[keyof typeof MenuCategory];

export const OrderStatus = {
  PENDING: "PENDING",
  CONFIRMED: "CONFIRMED",
  PREPARING: "PREPARING",
  READY: "READY",
  SERVED: "SERVED",
  COMPLETED: "COMPLETED",
  CANCELLED: "CANCELLED",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderType = {
  TAKEAWAY: "TAKEAWAY",
  DINING: "DINING",
  UBER: "UBER",
} as const;

export type OrderType = (typeof OrderType)[keyof typeof OrderType];